backend: Move functions from podutils to rokutils
