+++ backend/kale/tests/assets/functions/func03.out.py
@@ -2,15 +2,15 @@ def test():
     from kale.common import mlmdutils as _kale_mlmdutils
     _kale_mlmdutils.init_metadata()
 
-    from kale.common import podutils as _kale_podutils
+    from kale.common import rokutils as _kale_rokutils
     _kale_mlmdutils.call("link_input_rok_artifacts")
-    _kale_podutils.snapshot_pipeline_step(
+    _kale_rokutils.snapshot_pipeline_step(
         "test",
         "test",
         "/path/to/nb",
         before=True)
 
-    _rok_snapshot_task = _kale_podutils.snapshot_pipeline_step(
+    _rok_snapshot_task = _kale_rokutils.snapshot_pipeline_step(
         "test",
         "test",
         "/path/to/nb",
