+++ backend/kale/tests/assets/functions/func07.out.py
@@ -2,9 +2,9 @@ def final_auto_snapshot():
     from kale.common import mlmdutils as _kale_mlmdutils
     _kale_mlmdutils.init_metadata()
 
-    from kale.common import podutils as _kale_podutils
+    from kale.common import rokutils as _kale_rokutils
     _kale_mlmdutils.call("link_input_rok_artifacts")
-    _rok_snapshot_task = _kale_podutils.snapshot_pipeline_step(
+    _rok_snapshot_task = _kale_rokutils.snapshot_pipeline_step(
         "test",
         "final_auto_snapshot",
         "/path/to/nb",
